requirements.txt
google-cloud-storage == 2.14.0